Cloud handles pissing by pressing and releasing spacebar. While held down, Cloud will make WaterDrops. WaterDrops do not interact with other WaterDrops or with Cloud, but if they hit anything else, they are destroyed. If they hit a WaterReceiver, they call that object's ReceiveWater method.
A stand in flower has been added to my scene to show an example of this behavior. If water hits the flower, it gains 1 water level. It loses .2 waterlevel per second. If it gets to 0 or 15 it dies.

I created an interface, IWaterReceiver, to help simplify things on the water drop side.

I wanted to use Unity's particle system for water, but it doesn't interact with 2D objects.

Check the prefab folder for all the stuff I made. Grab the prefab and put it into your scene. That way changes will go across scenes with each commit.

Finally, linear drag has been added to Cloud to keep its movement a little more under control.